+++ src/main/java/com/example/ondongnae/backend/category/model/StoreSubCategory.java
@@ -1,5 +1,6 @@
-package com.example.ondongnae.backend.model;
+package com.example.ondongnae.backend.category.model;
 
+import com.example.ondongnae.backend.store.model.Store;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -13,7 +14,7 @@ public class StoreSubCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="sub_category_id")
