+++ src/main/java/com/example/ondongnae/backend/category/model/SubCategory.java
@@ -1,10 +1,9 @@
-package com.example.ondongnae.backend.model;
+package com.example.ondongnae.backend.category.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,25 +12,29 @@ import java.util.List;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MainCategory {
+public class SubCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @Column(nullable = false)
     private String name_ko;
+
     @Column(nullable = false)
     private String name_en;
+
     @Column(nullable = false)
     private String name_ja;
+
     @Column(nullable = false)
     private String name_zh;
 
-    @OneToMany(mappedBy = "mainCategory")
-    private List<Store> stores = new ArrayList<>();
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="main_category_id")
+    private MainCategory mainCategory;
 
-    @OneToMany(mappedBy = "mainCategory")
-    private List<SubCategory> subCategories = new ArrayList<>();
+    @OneToMany(mappedBy = "subCategory")
+    private List<StoreSubCategory> storeSubCategories = new ArrayList<>();
 
 }
