+++ src/main/java/com/example/ondongnae/backend/course/model/Course.java
@@ -1,4 +1,4 @@
-package com.example.ondongnae.backend.model;
+package com.example.ondongnae.backend.course.model;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -14,23 +14,29 @@ public class Course {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @Column(nullable = false)
     private String title_ko;
+
     @Column(nullable = false)
     private String title_en;
+
     @Column(nullable = false)
     private String title_zh;
+
     @Column(nullable = false)
     private String title_ja;
 
     @Column(nullable = false)
     private String description_ko;
+
     @Column(nullable = false)
     private String description_en;
+
     @Column(nullable = false)
     private String description_zh;
+
     @Column(nullable = false)
     private String description_ja;
